created the SQL test
the code copied from mySQL file on my computer if code is ready to be used can easily be edited to use the existing database or create a new one with a more suited name.
additionally the code here is all that is needed to create a basic database as a start.
TODO: we got to get the java working with the SQL database